% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{apply.fun.to.files}
\alias{apply.fun.to.files}
\title{Apply get function to a list of files}
\usage{
apply.fun.to.files(files, path, fun, ...)
}
\arguments{
\item{files}{the list of files to get value from}

\item{path}{the folder containing the list of files}

\item{fun}{function to apply to each file to retrieve the desired value,
examples include get.object.count.by.type, get.object.names,
get.field.value. The arguments of fun is in the "..."}
}
\value{
a dataframe with two columns, filename, and value
}
\description{
Aim to retrieve some statistics or field value for a group of idf files (in
the same folder) what stats/value to get depends on "fun" argument
}
