% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_output.R
\name{compare.before.after.col}
\alias{compare.before.after.col}
\title{Plot comparison between two eplusout columns}
\usage{
compare.before.after.col(
  df.before,
  df.after,
  col,
  ylabel,
  label.before = "before",
  label.after = "after",
  month = NULL,
  day = NULL
)
}
\arguments{
\item{df.before}{baseline case data frame, read in from eplusout.csv.
`Date/Time` data frame converted with something like this
read.idfEnergyPlus::convert.timestamp.eplusout(target.year = 2023)}

\item{df.after}{feature case data frame, read in from eplusout.csv. Need to
have `Date/Time` converted similarly as df.before}

\item{col}{column name to plot}

\item{ylabel}{yaxis label, something like "Temperature (C)"}

\item{month}{filter to a certain month}

\item{day}{filter to a certain day. The combination of the two facilitates
plotting of a typical summer day or winter day}
}
\value{
a ggplot2 object of line plot
}
\description{
Plot comparison between two eplusout columns
}
