% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_output.R
\name{convert.timestamp.eplusout}
\alias{convert.timestamp.eplusout}
\title{Convert timestamp}
\usage{
convert.timestamp.eplusout(df, target.year)
}
\arguments{
\item{df}{eplusout.csv directly}

\item{year}{year of the simulation result. EnergyPlus output doesn't have year in the timestamp}
}
\value{
a data frame with `Date/Time` converted to POSIXct
}
\description{
Convert timestamp
}
