% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{empty.or.with.end.marker}
\alias{empty.or.with.end.marker}
\title{Check whether a line is empty or has an end-of-object marker}
\usage{
empty.or.with.end.marker(line)
}
\arguments{
\item{line}{a line of the idf file}
}
\value{
TRUE if the line doesn't contain a comma (indicate the line is in the
    middle of an idf object), or contains a end-of-object marker";"
}
\description{
Check whether a line is empty or has an end-of-object marker
}
