% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.field.value.from.one.line}
\alias{get.field.value.from.one.line}
\title{get the value part from a line}
\usage{
get.field.value.from.one.line(field.line)
}
\arguments{
\item{field.lines}{a line whose value needs to be parsed}
}
\value{
the portion before the first "," or a ";"
}
\description{
The field value ends with either a "," or a ";". This function retrieves the part before a "," or ";" for a given line
}
