% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.field.value}
\alias{get.field.value}
\title{Get the value of an idf object field}
\usage{
get.field.value(
  lines,
  field.name,
  object.type = "",
  object.name = "",
  verbose = FALSE
)
}
\arguments{
\item{lines}{lines read from idf file}

\item{field.name}{name of the field whose value is to be retrieved}

\item{object.type}{the type of the object, like, "Construction,",
"Material,", etc. need to contain the "," as some object types are a
prefix of others}

\item{object.name}{name of the object. it is guarenteed unique from
EnergyPlus. If supply with empty string, then the name field is ignored}

\item{verbose}{if true, print debug outputs}
}
\value{
a string version of the value (might have prefix spaces)
}
\description{
Get the value of an idf object field
}
