% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.idf.chunk}
\alias{get.idf.chunk}
\title{Get the idf chunks given specified type or object name}
\usage{
get.idf.chunk(lines, object.type = "", object.name = "", verbose = FALSE)
}
\arguments{
\item{lines}{lines read from idf file}

\item{object.type}{the type of the object, like, "Construction,",
"Material,", etc. need to contain the "," as some object types are a
prefix of others}

\item{object.name}{name of the object. The default value is an empty string
The name can be a regex}
}
\value{
a vector of strings containing the chunks matching the specified
    object type and name
}
\description{
Get the idf chunks given specified type or object name
}
