% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.object.count.by.type}
\alias{get.object.count.by.type}
\title{Get number of objects}
\usage{
get.object.count.by.type(lines, object.type)
}
\arguments{
\item{lines}{lines read from idf file}

\item{object.type}{the type of the object, like, "Construction,",
"Material,", etc. need to contain the "," as some object types are a
prefix of others}
}
\value{
an integer of the number of object instances in the idf file
}
\description{
Get number of objects
}
