% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.object.end.idx}
\alias{get.object.end.idx}
\title{Get indices of idf object end}
\usage{
get.object.end.idx(lines)
}
\arguments{
\item{lines}{lines read from idf file}
}
\value{
a vector of integers, marking the line numbers of the end of all idd objects
}
\description{
Get indices of idf object end
}
