% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.object.line.idx.by.name}
\alias{get.object.line.idx.by.name}
\title{Get objects starting line indices by name
fixme: this cannot guarantee this line is the name line}
\usage{
get.object.line.idx.by.name(lines, object.name = "")
}
\arguments{
\item{lines}{lines read from idf file}

\item{object.name}{name of the object. The default value is an empty string.
The name can be a regex}
}
\value{
a vector of integers, marking the line numbers of the start of the
    object (like the location of "Construction,")
}
\description{
Get objects starting line indices by name
fixme: this cannot guarantee this line is the name line
}
