% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.object.line.idx.by.type}
\alias{get.object.line.idx.by.type}
\title{Get objects line indices}
\usage{
get.object.line.idx.by.type(lines, object.type = "")
}
\arguments{
\item{lines}{lines read from idf file}

\item{object.type}{the type of the object, like, "Construction,",
"Material,", etc. need to contain the "," as some object types are a
prefix of others. The default value is an empty string}
}
\value{
a vector of integers, marking the line numbers of the start of the object
}
\description{
Get objects line indices
}
