% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.object.line.idx}
\alias{get.object.line.idx}
\title{Get objects starting line indices by type and or name}
\usage{
get.object.line.idx(lines, object.type = "", object.name = "", verbose = FALSE)
}
\arguments{
\item{lines}{lines read from idf file}

\item{object.type}{the type of the object, like, "Construction,",
"Material,", etc. need to contain the "," as some object types are a
prefix of others. The default value is an empty string}

\item{object.name}{name of the object. The default value is an empty string
The name can be a regex}
}
\value{
a vector of integers, marking the line numbers of the start of the
    object (like the location of "Construction,")
}
\description{
Get objects starting line indices by type and or name
}
