% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.object.names}
\alias{get.object.names}
\title{Get names of objects for a certain type}
\usage{
get.object.names(lines, object.type)
}
\arguments{
\item{lines}{lines read from idf file}

\item{object.type}{the type of the object, like, "Construction,",
"Material,", etc. need to contain the "," as some object types are a
prefix of others}
}
\value{
a vector of strings containing the object name lines for the
    specified object type
}
\description{
Get names of objects for a certain type
}
