% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{get.start.end.idx}
\alias{get.start.end.idx}
\title{Get start and end line index of idf objects}
\usage{
get.start.end.idx(lines, object.type = "", object.name = "", verbose = FALSE)
}
\arguments{
\item{lines}{lines read from idf file}

\item{object.type}{the type of the object, like, "Construction,",
"Material,", etc. need to contain the "," as some object types are a
prefix of others}

\item{object.name}{name of the object. The default value is an empty string
The name can be a regex}
}
\value{
two equal-length vectors marking the start and end of the objects
}
\description{
Get start and end line index of idf objects
}
