% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{read.schedule}
\alias{read.schedule}
\title{Read a schedule object into a data frame}
\usage{
read.schedule(lines, schedule.type, schedule.name, verbose = FALSE)
}
\arguments{
\item{schedule.type}{first line of a schedule object, like "Schedule:Day:Interval,"}

\item{schedule.name}{object name, like "single_family_first_story heating week 0 Sunday,"}
}
\value{
a dataframe with two columns with string type, "time" and "value". The "time" column contains the value of the schedule.
}
\description{
Read a schedule object into a data frame
}
