% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{replace.design.day}
\alias{replace.design.day}
\title{replace design day object in an input idf with a corresponding design day
object in a ddy file}
\usage{
replace.design.day(lines, ddy.lines, design.day.kw, verbose = FALSE)
}
\arguments{
\item{lines}{lines read from idf file}

\item{ddy.lines}{lines read from ddy file containing design days}

\item{design.day.kw}{design day object name or a substring (can be regex) to
locate the design day object in the idd file and to match the
corresponding one in the ddy file}

\item{verbose}{if true, print debug outputs}
}
\value{
a vector of lines with the chunk replaced
}
\description{
replace design day object in an input idf with a corresponding design day
object in a ddy file
}
