% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{replace.field.value}
\alias{replace.field.value}
\title{replace the value of a field in idf file}
\usage{
replace.field.value(
  lines,
  object.type,
  object.name,
  field.name,
  fun,
  verbose = FALSE
)
}
\arguments{
\item{lines}{lines read from idf file}

\item{object.type}{the type of the object, like, "Construction,",
"Material,", etc. need to contain the "," as some object types are a
prefix of others}

\item{object.name}{name of the object. it is guarenteed unique from
EnergyPlus. If this is not specified and object.type have multiple
instances, then only the first chunk is replaced}

\item{field.name}{name of the field whose value is to be retrieved}

\item{fun}{if the new value is a function of the old value, pass the function
here. To replace it with a specific value, pass in a constant function.
The function input is a string}

\item{verbose}{if true, print debug outputs}

\item{new.value}{the text used to substitute the old value}
}
\value{
a vector of lines with the value replaced
}
\description{
replace value of a field in idf file
}
