% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_idf.R
\name{replace.idf.chunk}
\alias{replace.idf.chunk}
\title{replace a chunk (object) in idf file}
\usage{
replace.idf.chunk(
  lines,
  new.text,
  object.type = "",
  object.name = "",
  verbose = FALSE
)
}
\arguments{
\item{lines}{lines read from idf file}

\item{new.text}{the text used to substitute the old chunk, starting from the
object name line}

\item{object.type}{the type of the object, like, "Construction,",
"Material,", etc. need to contain the "," as some object types are a
prefix of others}

\item{object.name}{name of the object. it is guarenteed unique from
EnergyPlus. If this is not specified and object.type have multiple
instances, then only the first chunk is replaced}

\item{verbose}{if true, print debug outputs}
}
\value{
a vector of lines with the chunk replaced
}
\description{
This searches the start of the idf object by substring.pattern, and replace
this chunk with a new chunk
}
